Add exit/restart functionality
Added the ability to invoke a slideshow on demand instead of
automatically creating it on page load. This is useful when you have a
normal web page with a 'start slide show' button that should invoke the
plug-in. Added an exit button to the shutter theme to allow for
returning to the web page. Added demo3.html to demonstrate the feature.