Added options 'html_template' and 'auto_exit'; changed loop option
Added option 'html_template' so that the HTML control template can be
passed as a variable. Added option 'auto_exit' that, when true (1), the
slideshow will automatically exit after the last slide is shown. Renamed
'stop_loop' option to 'loop' and changed functionality so that, when
true (1), navigation is allowed between the first and last slide. When
false (0), no action is taken when user tries to navigate beyond the
first and last slides. Changed jQuery reference to local version for
faster page loading.