BUG FIX: Fix for Chrome Mobile
On Chrome Mobile (Nexus 7), images are not visible (just a dark
background). The fix was to set visibility:hidden on the ul element
containing the slides and to make sure the first slide was set to
visibility:visible. This is a bit counter-intuitive and I can't explain
why it works, but it does without having any apparent negative
consequences on other browsers.