Hide progress bar & move slide bullets
When the title is long or a user's screen is small, the title and slide
bullet list can render on top of each other. To minimize this, I made
two changes (1) Right align the bullet list (2) Use a CSS media query to
show the list only for larger screens (> 800px wide). I also forced the
progress bar to be completely invisible when it is disabled
(progress_bar=0).