BUG FIX: Slide bullets cover buttons
When there are a lot of slide bullets, they can cover the title and the
pause/exit buttons on the left. I changed it so a max-width is assigned
to their container to prevent this. Note that the bullets may still
cover a portion of the title when it is long, but it was not practical
to fix it.