BUG FIX: Buttons don't work for single image slide show
The plug-in detects when there is a single image in the slide show and
sets base.options.slideshow to 0 in this case. However, this causes the
pause and cancel buttons to stop working. I commented out this line so
that behavior for a single image slide show is similar to one with
multiple images.