Cancel resize event in destroy()
The widget adds two resize event handlers on 'window' that were never unsubscribed, causing them to continue running after a slideshow has ended and the user resizes the window. Since at that point the api and vars variables were null, a JavaScript error occurred. The fix was to unsubscribe from the resize events in the destroy function.